// src/app/models/requests/category/update-category-request.ts
export class UpdateCategoryRequest {
    id: string = '';
    name: string = '';
    codeCategory: string = '';
    parentCategoryId?: string;

    validationMessages = {
        "name": {
            required: "Tên loại sản phẩm là bắt buộc."
        },
        "codeCategory": {
            required: "Mã loại sản phẩm là bắt buộc."
        },
    }
}